refactor: remove store providers and refactor playground
