+++ packages/presets/src/editors/index.ts
@@ -1,6 +1,3 @@
-import './editor-container.js';
-
-export * from './affine-doc-editor.js';
 export * from './doc-editor.js';
 export * from './edgeless-editor.js';
 export * from './editor-container.js';
